+++ dart/lib/src/protocol/sentry_user.dart
@@ -1,5 +1,7 @@
 import 'package:meta/meta.dart';
 
+import '../../sentry.dart';
+
 /// Describes the current user associated with the application, such as the
 /// currently signed in user.
 ///
@@ -37,12 +39,20 @@ class SentryUser {
     this.email,
     this.ipAddress,
     this.segment,
-    Map<String, dynamic>? extras,
-  })  : assert(id != null ||
-            username != null ||
-            email != null ||
-            ipAddress != null ||
-            segment != null),
+    this.geo,
+    this.name,
+    Map<String, dynamic>? data,
+    @Deprecated('Will be removed in v7. Use [data] instead')
+        Map<String, dynamic>? extras,
+  })  : assert(
+          id != null ||
+              username != null ||
+              email != null ||
+              ipAddress != null ||
+              segment != null,
+        ),
+        data = data == null ? null : Map.from(data),
+        // ignore: deprecated_member_use_from_same_package
         extras = extras == null ? null : Map.from(extras);
 
   /// A unique identifier of the user.
@@ -64,13 +74,38 @@ class SentryUser {
   ///
   /// These keys are stored as extra information but not specifically processed
   /// by Sentry.
+  final Map<String, dynamic>? data;
+
+  @Deprecated('Will be removed in v7. Use [data] instead')
   final Map<String, dynamic>? extras;
+
+  /// Approximate geographical location of the end user or device.
+  ///
+  /// The geolocation is automatically inferred by Sentry.io if the [ipAddress] is set.
+  /// Sentry however doesn't collect the [ipAddress] automatically because it is PII.
+  /// The geo location will currently not be synced to the native layer, if available.
+  // See https://github.com/getsentry/sentry-dart/issues/1065
+  final SentryGeo? geo;
+
+  /// Human readable name of the user.
+  final String? name;
 
   /// Deserializes a [SentryUser] from JSON [Map].
   factory SentryUser.fromJson(Map<String, dynamic> json) {
     var extras = json['extras'];
     if (extras != null) {
-      extras = Map<String, dynamic>.from(extras as Map);
+      extras = Map<String, dynamic>.from(extras);
+    }
+
+    var data = json['data'];
+    if (data != null) {
+      data = Map<String, dynamic>.from(data);
+    }
+
+    SentryGeo? geo;
+    final geoJson = json['geo'];
+    if (geoJson != null) {
+      geo = SentryGeo.fromJson(Map<String, dynamic>.from(geoJson));
     }
     return SentryUser(
       id: json['id'],
@@ -78,19 +113,28 @@ class SentryUser {
       email: json['email'],
       ipAddress: json['ip_address'],
       segment: json['segment'],
+      data: data,
+      geo: geo,
+      name: json['name'],
+      // ignore: deprecated_member_use_from_same_package
       extras: extras,
     );
   }
 
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
+    final geoJson = geo?.toJson();
     return <String, dynamic>{
       if (id != null) 'id': id,
       if (username != null) 'username': username,
       if (email != null) 'email': email,
       if (ipAddress != null) 'ip_address': ipAddress,
       if (segment != null) 'segment': segment,
+      if (data?.isNotEmpty ?? false) 'data': data,
+      // ignore: deprecated_member_use_from_same_package
       if (extras?.isNotEmpty ?? false) 'extras': extras,
+      if (name != null) 'name': name,
+      if (geoJson != null && geoJson.isNotEmpty) 'geo': geoJson,
     };
   }
 
@@ -100,14 +144,23 @@ class SentryUser {
     String? email,
     String? ipAddress,
     String? segment,
-    Map<String, dynamic>? extras,
-  }) =>
-      SentryUser(
-        id: id ?? this.id,
-        username: username ?? this.username,
-        email: email ?? this.email,
-        ipAddress: ipAddress ?? this.ipAddress,
-        segment: segment ?? this.segment,
-        extras: extras ?? this.extras,
-      );
+    @Deprecated('Will be removed in v7. Use [data] instead')
+        Map<String, dynamic>? extras,
+    String? name,
+    SentryGeo? geo,
+    Map<String, dynamic>? data,
+  }) {
+    return SentryUser(
+      id: id ?? this.id,
+      username: username ?? this.username,
+      email: email ?? this.email,
+      ipAddress: ipAddress ?? this.ipAddress,
+      segment: segment ?? this.segment,
+      data: data ?? this.data,
+      // ignore: deprecated_member_use_from_same_package
+      extras: extras ?? this.extras,
+      geo: geo ?? this.geo,
+      name: name ?? this.name,
+    );
+  }
 }