+++ flutter/test/sentry_flutter_test.dart
@@ -4,11 +4,14 @@ import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/debug_print_integration.dart';
 import 'package:sentry_flutter/src/version.dart';
 import 'mocks.dart';
+import 'mocks.mocks.dart';
 import 'sentry_flutter_util.dart';
 
 /// These are the integrations which should be added on every platform.
 /// They don't depend on the underlying platform.
 final platformAgnosticIntegrations = [
+  WidgetsFlutterBindingIntegration,
+  OnErrorIntegration,
   FlutterErrorIntegration,
   LoadReleaseIntegration,
   DebugPrintIntegration,
@@ -39,111 +42,207 @@ void main() {
     });
 
     test('Android', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: true,
-          shouldHaveIntegrations: [
-            ...androidIntegrations,
-            ...nativeIntegrations,
-            ...platformAgnosticIntegrations,
-          ],
-          shouldNotHaveIntegrations: iOsAndMacOsIntegrations,
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(platform: MockPlatform.android()),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: true,
+      );
+
+      testConfiguration(
+          integrations: integrations,
+          shouldHaveIntegrations: [
+            ...androidIntegrations,
+            ...nativeIntegrations,
+            ...platformAgnosticIntegrations,
+          ],
+          shouldNotHaveIntegrations: iOsAndMacOsIntegrations);
+
+      integrations
+          .indexWhere((element) => element is WidgetsFlutterBindingIntegration);
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     }, testOn: 'vm');
 
     test('iOS', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: true,
-          shouldHaveIntegrations: [
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-            ...platformAgnosticIntegrations,
-          ],
-          shouldNotHaveIntegrations: androidIntegrations,
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(platform: MockPlatform.iOs()),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: true,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: [
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+          ...platformAgnosticIntegrations,
+        ],
+        shouldNotHaveIntegrations: androidIntegrations,
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     }, testOn: 'vm');
 
     test('macOS', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: true,
-          shouldHaveIntegrations: [
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-            ...platformAgnosticIntegrations,
-          ],
-          shouldNotHaveIntegrations: androidIntegrations,
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(platform: MockPlatform.macOs()),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: true,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: [
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+          ...platformAgnosticIntegrations,
+        ],
+        shouldNotHaveIntegrations: androidIntegrations,
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     }, testOn: 'vm');
 
     test('Windows', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(platform: MockPlatform.windows()),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     }, testOn: 'vm');
 
     test('Linux', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(platform: MockPlatform.linux()),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     }, testOn: 'vm');
 
     test('Web', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(
@@ -152,22 +251,41 @@ void main() {
         ),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     });
 
     test('Web && (iOS || macOS) ', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       // Tests that iOS || macOS integrations aren't added on a browser which
       // runs on iOS or macOS
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(
@@ -176,22 +294,41 @@ void main() {
         ),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     });
 
     test('Web && (macOS)', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       // Tests that iOS || macOS integrations aren't added on a browswer which
       // runs on iOS or macOS
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(
@@ -200,21 +337,40 @@ void main() {
         ),
       );
 
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
+
       await Sentry.close();
     });
 
     test('Web && Android', () async {
+      List<Integration> integrations = [];
+      Transport transport = MockTransport();
+
       // Tests that Android integrations aren't added on an Android browswer
       await SentryFlutter.init(
-        getConfigurationTester(
-          hasFileSystemTransport: false,
-          shouldHaveIntegrations: platformAgnosticIntegrations,
-          shouldNotHaveIntegrations: [
-            ...androidIntegrations,
-            ...iOsAndMacOsIntegrations,
-            ...nativeIntegrations,
-          ],
-        ),
+        (options) async {
+          options.dsn = fakeDsn;
+          integrations = options.integrations;
+          transport = options.transport;
+        },
         appRunner: appRunner,
         packageLoader: loadTestPackage,
         platformChecker: getPlatformChecker(
@@ -222,6 +378,26 @@ void main() {
           platform: MockPlatform.android(),
         ),
       );
+
+      testTransport(
+        transport: transport,
+        hasFileSystemTransport: false,
+      );
+
+      testConfiguration(
+        integrations: integrations,
+        shouldHaveIntegrations: platformAgnosticIntegrations,
+        shouldNotHaveIntegrations: [
+          ...androidIntegrations,
+          ...iOsAndMacOsIntegrations,
+          ...nativeIntegrations,
+        ],
+      );
+
+      testBefore(
+          integrations: integrations,
+          beforeIntegration: WidgetsFlutterBindingIntegration,
+          afterIntegration: OnErrorIntegration);
 
       await Sentry.close();
     });